requirements.txt
accelerate==0.33.0
albumentations==1.4.14
copairs==0.0.3
efaar_benchmarking==1.0
ftfy==6.3.1
h5py==3.12.1
huggingface_hub==0.26.3
numpy==1.24.4
openai==1.56.0
opencv_python_headless==4.10.0.84
packaging==24.2
pandas==2.0.3
Pillow==11.0.0
rdkit_pypi==2022.3.4
regex==2024.7.24
scikit_learn==1.5.2
scipy==1.14.1
timm==1.0.11
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.5
training==0.0.1
transformers==4.46.3
git+https://github.com/cytomining/copairs@880f22a551bd897896d148a0b07baa99d981c6a9
git+https://github.com/ml-jku/hopfield-layers
